fix(schedules): cast 0041's day_count, regenerate schema.ts, break the client cycle
Closes #537, closes #538.

- 0041's v_training_schedules.day_count gains ::int: count(*) is bigint,
  which node-postgres returns as a string, against dayCount: number. Edited
  in place under D344 — unmerged, and db:status/db:status:prod both reported
  it pending. Applied to Neon dev afterwards; 0041's verification script ran
  for the first time, 9/9 PASS.
- schema.ts regenerated by drizzle-kit introspect against the migrated
  database, clearing the four TS2305 errors that stopped CI's type gate.
  The real view types then caught findScheduleIdsUsingRoutine returning
  (string | null)[] — Drizzle types every view column nullable.
- unwrapOrThrow moves from client.ts to unwrap.ts (D346): client.ts →
  sessions.ts → client.ts was a cycle the fallow gate rejects, which CI
  never reached because the type gate failed first.
- schema-view-drift normalizes two Postgres echo artifacts confirmed with
  pg_get_viewdef: an aggregate's implicit output alias, and the qualifier
  drop in a single-relation select list (D345).
- toThrowError → toThrow; the deprecation hint fails astro check.

Docs: 03-Migrations.md and 02-Template-Layer.md record the cast and the
verification run, CLAUDE.md's migration invariant is rewritten around
D344's proof rule, File Inventory gains unwrap.ts, history 1.109.0.

+++ database/migrations/0041_training_schedules.sql
@@ -55,7 +55,7 @@ SELECT ts.id AS schedule_id,
     ts.name,
     ts.is_active,
     ts.updated_at,
-    (SELECT count(*) FROM training_schedule_days d WHERE d.training_schedule_id = ts.id) AS day_count
+    (SELECT count(*) FROM training_schedule_days d WHERE d.training_schedule_id = ts.id)::int AS day_count
 FROM training_schedules ts;
 COMMENT ON VIEW v_training_schedules IS 'One row per schedule with its day count; filter by player_id.';
